Following the structure of the Room codelab
- Makes GameScreen the owner of GameScreenViewModel
- Use a viewModelFactory
- Use a custom Application
- Created a general repository for curent session, games history and a general accessor for the backend
- Created offline database implementations of the repositories
- Created an accessor to the online backend
- Created AppContainer, a class that will hold all necessary singletons for the app
- Generalized and refactored GameScreen.kt
- Moved things around

The main motivation behind these changes (refactoring of the view aside) was to make way for possible enhancements.
I have been asked to maybe have a "Random stop" mode, which would need an offline gameplay mode. I'd need the same methods to interact with a backend (running locally) than i do with OnlineGameInteraction.
Same for saving user metrics. I do that offline in a local database, but if i want to do the same with a cloud database I can just make an OnlineBlahBlahRepository that implements the correct Repository.

This isn't very YAGNI, I admit.